# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const SignupBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
});

export const SignupResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Log in and receive a JWT token
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.string(),
    name: zod.string(),
    email: zod.string(),
  }),
});

/**
 * @summary Get all tasks for the logged-in user
 */
export const GetTasksResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  createdAt: zod.string(),
});
export const GetTasksResponse = zod.array(GetTasksResponseItem);

/**
 * @summary Add a new task
 */
export const AddTaskBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteTaskResponse = zod.object({
  success: zod.boolean(),
});